+++ changelog.md
@@ -1,5 +1,9 @@
 #### UPDATED:
-- Fixed an issue with quark module
+- **Macaw's Door** (COMMON) - Fixed _bark_glass_door_'s item texture
+- **Macaw's Fences & Walls** (FORGE) - Related to **The Aether: Redux** & **The Aether** - Fixed _hedge_'s texturing problems 
+
+#### OTHERS: 
+- **FPW Aesthetic Gems** (COMMMON) - Fixed the 2 woods' texturing problems - Related to **Macaw's Furniture**
 
 ---
 **LEGEND**: